+++ lib/src/groups/dh_group.dart
@@ -25,12 +25,17 @@ enum DhGroup {
   g17,
 
   /// 8192-bit MODP Group.
-  g18;
+  g18,
 
-  /// Get the Diffie-Hellman parameters for this group
-  /// If [privateValueLength] is not provided, the default value of 2048 is used.
-  DhParameter getParameter({int? privateValueLength}) => getDhParameter(
-        this,
-        privateValueLength: privateValueLength,
-      );
+  /// 1024-bit MODP Group with 160-bit prime order subgroup.
+  g22,
+
+  /// 2048-bit MODP Group with 224-bit prime order subgroup.
+  g23,
+
+  /// 2048-bit MODP Group with 256-bit prime order subgroup.
+  g24;
+
+  /// The Diffie-Hellman parameters for this group
+  DhParameter get parameter => getParameter(this);
 }